@@ -2,41 +2,41 @@
     "name": "kjudge-frontend",
     "version": "0.7.0",
     "description": "Templates and others for kjudge.",
-    "main": "index.js",
     "repository": "https://github.com/natsukagami/kjudge",
     "author": "Natsu Kagami",
     "license": "AGPL-3.0",
     "private": true,
-    "devDependencies": {
-        "@fullhuman/postcss-purgecss": "^3.1.3",
-        "@tailwindcss/postcss7-compat": "^2.0.2",
-        "@types/humanize-duration": "^3.18.0",
-        "@types/node": "^14.14.14",
-        "@types/react": "^17.0.0",
-        "autoprefixer": "^9.7.5",
-        "fontsource-ibm-plex-mono": "^3.1.5",
-        "fontsource-muli": "^3.1.5",
-        "parcel-bundler": "^1.12.4",
-        "prettier": "2.2.1",
-        "tailwindcss": "npm:@tailwindcss/postcss7-compat",
-        "typescript": "^4.1.3"
+    "devDependencies": {        
+        "@fullhuman/postcss-purgecss": "5.0.0",
+        "@types/humanize-duration": "3.27.1",
+        "@types/node": "20.1.3",
+        "@types/react": "18.2.6",
+        "autoprefixer": "10.4.14",
+        "parcel": "2.8.3",
+        "postcss": "8.4.23",
+        "prettier": "2.8.8",
+        "tailwindcss": "3.3.2",
+        "typescript": "5.0.4"
     },
     "scripts": {
         "test": "echo No tests... yet",
         "fmt": "prettier \"css/**/*\" \"ts/**/*\" \"*.json\" \"*.js\"",
-        "dev": "parcel watch -d ../embed/templates \"html/**/*.html\"",
-        "build": "mkdir -p ../embed/templates && rm -rf ../embed/templates/* && parcel build --no-source-maps -d ../embed/templates \"html/**/*.html\""
+        "dev": "parcel watch \"html/**/*.html\" --dist-dir ../embed/templates",
+        "build": "parcel build \"html/**/*.html\" --no-source-maps --dist-dir ../embed/templates"
     },
     "alias": {
         "react": "preact/compat",
         "react-dom/test-utils": "preact/test-utils",
-        "react-dom": "preact/compat"
+        "react-dom": "preact/compat",
+        "preact/jsx-dev-runtime": "preact/jsx-runtime"
     },
     "dependencies": {
-        "highlight.js": "^10.4.1",
-        "humanize-duration": "^3.22.0",
-        "preact": "^10.3.4",
-        "react-flip-move": "^3.0.4",
-        "regenerator-runtime": "^0.13.5"
+        "@fontsource/ibm-plex-mono": "4.5.13",
+        "@fontsource/mulish": "4.5.14",
+        "highlight.js": "11.8.0",
+        "humanize-duration": "3.28.0",
+        "preact": "10.13.2",
+        "react-flip-move": "3.0.5",
+        "regenerator-runtime": "0.13.11"
     }
 }
